ECCW-404: Patch cludo interface to use custom template
The Cludo instance for ECC has been configured with a custom template, something that cludo now is loathe to do for clients. The custom template is selected by omitting the type selection, rather than selecting one of the two built-in templates. This patch removes that field from the API calls.