ECCW-410: Allow chmod to fail in some circumstances
The Drupal core uses a wrapped chmod function that handles getting the correct modes. This is used in multiple places, but especially the image scale generation code. This function assumes a well-behaved filesystem that allows chmod, but Azure blob storage doesn't. This patch changes the failure path of the function to report success if the file has a mode of 777, even if the chmod fails. This is needed as Azure limits the mode to only being this single allowed value when mounting a blob storage in a container app. The result is that the initial image scale call succeeds and returns the file, rather than succeeding and returning an error.